requirements.txt
# requirements.txt (V-Final for Streamlit Cloud)

# Core Framework
streamlit>=1.30.0

# AI & Language Models
google-generativeai>=0.4.0

# Data Sources & Connectors
pinecone-client>=3.0.0
neo4j>=5.17.0

# Utilities
PyYAML>=6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyMuPDF>=1.22.0
# Required for vector embedding
sentence-transformers